Don't create special gerrit refs
Previously for supporting gerrit the special
refs/for/* and refs/drafts/* where created
in the request namespace so that git will see
them and use them to determine the "old" oid
for the update-hook that will then be used in the
revwalk.

Now the old oid is determined by the hook itself
taking into account the special refs.

This is needed to support pushes based on older
common ancestors.