[url] Switch to 'url' crate, as correctness certainly is more important than compile times
In the moment percent encoding is applied, it's impossible to not
allocate anymore, at least it's hard and not worth the effort given
how often a URL is actually used.

Let's hide the actual parser, and validate that the URLs git is
interested in are working and easy to use.