Remove self refernce link
The link actually points to exactly the some location, as reader
it does not make any sense to follow this links it is rather
anoying to find out if you open it in another tab that it
is the exact same page at the exact same position. There
is no benefit in having this link.