Greatly reduce compound::Object size
It costs an allocation in the uncommon case, so shouldn't be an issue.
Note that the object is constructed on the stack and the moved to the
heap, so there is still some efficiency to be gained. However,
that will probably never be relevant given that packs are zero copy
and the builk of objects live there.